Add stateChanges and tasks packages, add TODO's for javadocs
