+++ engine/src/main/java/org/terasology/rendering/dag/StateChange.java
@@ -16,7 +16,7 @@
 package org.terasology.rendering.dag;
 
 /**
- *
+ * TODO: Add javadocs
  */
 public interface StateChange<T> {
     T getValue();
